// src/types/register.ts
export interface RegisterFormData {
  teamName: string;
  college: string;
  leaderName: string;
  email: string;
  phone: string;

  department: string;
  year: number;

  teamSize: string;
  members: string;

  projectName: string;
  domain: string;
  problemStatement: string;
  solution: string;
  techStack: string;
  github: string;

  transactionId: string;
}